agregado interface, declarados precios extras

+++ src/main/java/TicketGenerator.java
@@ -1,5 +1,6 @@
 import product.Product;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -14,6 +15,6 @@ public class TicketGenerator {
     public Double getTotalPrice() {
         return products.stream()
                 .map(Product::getPrice)
-                .reduce(0.0, Double::sum);
+                .reduce(0.0, BigDecimal::sum);
     }
 }
